+++ src/components/Proposal/Create/SchemeForms/CreateUnknownGenericSchemeProposal.tsx
@@ -5,8 +5,9 @@ import { ErrorMessage, Field, Form, Formik, FormikProps } from "formik";
 import Analytics from "lib/analytics";
 import * as React from "react";
 import { connect } from "react-redux";
-import { showNotification } from "reducers/notifications";
+import { showNotification, NotificationStatus } from "reducers/notifications";
 import { isValidUrl } from "lib/util";
+import { exportUrl, importUrlValues } from "lib/proposalUtils";
 import TagsSelector from "components/Proposal/Create/SchemeForms/TagsSelector";
 import TrainingTooltip from "components/Shared/TrainingTooltip";
 import * as css from "../CreateProposal.scss";
@@ -40,16 +41,27 @@ interface IFormValues {
   title: string;
   url: string;
   value: number;
+  [key: string]: any;
 }
 
 class CreateGenericScheme extends React.Component<IProps, IStateProps> {
+
+  initialFormValues: IFormValues;
 
   constructor(props: IProps) {
     super(props);
 
     this.handleSubmit = this.handleSubmit.bind(this);
+    this.initialFormValues = importUrlValues<IFormValues>({
+      description: "",
+      callData: "",
+      title: "",
+      url: "",
+      value: 0,
+      tags: [],
+    });
     this.state = {
-      tags: new Array<string>(),
+      tags: this.initialFormValues.tags,
     };
   }
 
@@ -75,6 +87,12 @@ class CreateGenericScheme extends React.Component<IProps, IStateProps> {
     this.props.handleClose();
   }
 
+  // Exports data from form to a shareable url.
+  public exportFormValues(values: IFormValues) {
+    exportUrl({ ...values, ...this.state });
+    this.props.showNotification(NotificationStatus.Success, "Exportable url is now in clipboard :)");
+  }
+
   private onTagsChange = (tags: any[]): void => {
     this.setState({tags});
   }
@@ -88,12 +106,7 @@ class CreateGenericScheme extends React.Component<IProps, IStateProps> {
       <div className={css.contributionReward}>
         <Formik
           // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
-          initialValues={{
-            callData: "",
-            title: "",
-            url: "",
-            value: 0,
-          } as IFormValues}
+          initialValues={this.initialFormValues}
           // eslint-disable-next-line react/jsx-no-bind
           validate={(values: IFormValues): void => {
             const errors: any = {};
@@ -146,6 +159,7 @@ class CreateGenericScheme extends React.Component<IProps, IStateProps> {
             // eslint-disable-next-line @typescript-eslint/no-unused-vars
             setFieldTouched,
             setFieldValue,
+            values,
           }: FormikProps<IFormValues>) =>
             <Form noValidate>
               <TrainingTooltip overlay="The title is the header of the proposal card and will be the first visible information about your proposal" placement="right">
@@ -242,6 +256,11 @@ class CreateGenericScheme extends React.Component<IProps, IStateProps> {
               </div>
 
               <div className={css.createProposalActions}>
+                <TrainingTooltip overlay="Export proposal" placement="top">
+                  <button id="export-proposal" className={css.exportProposal} type="button" onClick={() => this.exportFormValues(values)}>
+                    <img src="/assets/images/Icon/share-blue.svg" />
+                  </button>
+                </TrainingTooltip>
                 <button className={css.exitProposalCreation} type="button" onClick={handleClose}>
                   Cancel
                 </button>