+++ src/components/Redemptions/RedemptionsPage.tsx
@@ -7,7 +7,7 @@ import Loading from "components/Shared/Loading";
 import withSubscription, { ISubscriptionProps } from "components/Shared/withSubscription";
 import gql from "graphql-tag";
 import Analytics from "lib/analytics";
-import { formatTokens, tokenSymbol } from "lib/util";
+import { formatTokens, tokenDecimals, tokenSymbol } from "lib/util";
 import { Page } from "pages";
 import * as React from "react";
 import { BreadcrumbsItem } from "react-breadcrumbs-dynamic";
@@ -202,7 +202,7 @@ class RedemptionsPage extends React.Component<IProps, null> {
       totalRewards.push(formatTokens(genReward, "GEN"));
     }
     Object.keys(externalTokenRewards).forEach((tokenAddress) => {
-      totalRewards.push(formatTokens(externalTokenRewards[tokenAddress], tokenSymbol(tokenAddress)));
+      totalRewards.push(formatTokens(externalTokenRewards[tokenAddress], tokenSymbol(tokenAddress), tokenDecimals(tokenAddress)));
     });
     if (reputationRewardDaos.size > 0) {
       totalRewards.push(<span>reputation in {reputationRewardDaos.size}&nbsp;DAOs</span>);