+++ CHANGELOG.md
@@ -4,7 +4,16 @@
 
 ## Next release
   - Features Added
+    - Improve performance of 3box actions like follow items for feed.
+    - Add informational popup before any action that interacts with 3box.
+    - Add follow button to DAO sidebar
+    - Improve copy of home page when not logged in
+    - Improve and clean up header text for feed items
   - Bugs Fixed
+    - Fix bug where tooltips in sidebar would cover other menu items. Move them to the right of the sidebar
+    - Log out of 3box when clicking the Log Out button
+    - Fix styling of Feed items
+
 
 ## 2019-12-18
 
@@ -17,9 +26,9 @@
 
 ## 2019-12-10
   - Features Added
-    - Use 3box profiles 
+    - Use 3box profiles
     - Create a Feed of events
-    - add new cookie policy and privacy policy 
+    - add new cookie policy and privacy policy
   - Bugs Fixed
     - do not error on bad token address
     - convert deprecated React methods
@@ -27,7 +36,7 @@
     - add subscriptions for votes, stakes and rewards on the proposal page
     - ignore externalTokenReward when address is null
     - change old dai to sai, add new dai
-    - better position notifications 
+    - better position notifications
     - shorten token dropdown menu
 
 ## 2019-12-03